+++ src/transaction_signer.js
